Working architecture for the project.
GameActivity contains one GridLayout. The activity dynamically creates
the create depending on the data given by Grid class. User input is also
taken care of.

Landscape orientation doesn't work properly yet.

An observer pattern coupled with a visitor pattern was used to
communicate between the view (GameActivity) and the Grid. A few classes
are given to identify which events are triggered by the Grid (events
package).